# services/control-plane-api/openapi/control_plane_openapi_v1.4.yaml
openapi: 3.0.3
info:
  title: AI Digital Office OS Control Plane
  version: 1.4.0
  description: >
    Governed control-plane API for AI DIGITAL OFFICE OS. Every endpoint is
    tenant-scoped via the JWT 'tenant_id' claim, versioned under /v1, and
    returns the standard error envelope on any non-2xx response (see
    ErrorResponse schema). v1.4 closes the integrity-review findings against
    v1.3: full project/agent-version/run/deployment/usage/tool lifecycle
    coverage, an explicit auth contract (issuer/audience/JWKS/scopes), and
    a required Idempotency-Key header on every state-changing endpoint.
    This specification is product-agnostic: no endpoint, schema, or example
    is specific to any named downstream project (see blueprint clause 75).
servers:
  - url: https://api.digitaloffice.example/v1
    description: Production (placeholder host — real endpoint is an
      environment-specific choice, intentionally not fabricated here)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2 client-credentials token for service/agent callers, or
        OAuth2 authorization-code token for human users (blueprint clause 70).
        Required claims: 'iss' (issuer — this Office's or an approved
        external IdP), 'aud' (this control plane), 'exp'/'iat' (max lifetime
        15 minutes for service/agent callers, 8 hours for human sessions),
        'tenant_id', 'scope' (granted permission scopes), and 'actor_type'
        (SERVICE | HUMAN). Clock-skew tolerance ±60 seconds. JWKS rotates on
        a maximum 90-day cycle with a minimum 7-day overlap window. Session
        and refresh-token lifecycle is delegated entirely to the issuing IdP
        (blueprint clause 71) — this API only validates presented tokens.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >
        Required on every state-changing (POST/PATCH/DELETE) request.
        Replaying the same key within 24 hours returns the original result
        rather than reprocessing (blueprint clauses 49, 63, 68).
  schemas:
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id, retryable]
      properties:
        error_code:
          type: string
          enum: [VALIDATION_ERROR, AUTHORIZATION_ERROR, POLICY_ERROR, TOOL_ERROR,
                 MODEL_ERROR, PROVIDER_ERROR, NETWORK_ERROR, TIMEOUT,
                 DATABASE_ERROR, ARTIFACT_ERROR, DEPLOYMENT_ERROR, UNKNOWN_ERROR]
        message: { type: string }
        correlation_id: { type: string }
        retryable: { type: boolean }
...
    AgentSpecification:
      type: object
      required: [agent_name, department, role, purpose, capabilities]
      properties:
        agent_name: { type: string }
        department: { type: string }
        role: { type: string }
        purpose: { type: string }
        capabilities: { type: array, items: { type: string } }
        required_model_capabilities: { type: array, items: { type: string } }
        allowed_tools: { type: array, items: { type: string } }
        permissions: { type: array, items: { type: string } }
        data_access: { type: array, items: { type: string } }
        input_schema: { type: object }
        output_schema: { type: object }
        security_level: { type: string, enum: [GREEN, YELLOW, RED], default: GREEN }
        sandbox_policy: { type: object }
        parent_agent_id: { type: string }
    AgentRecord:
      allOf:
        - $ref: '#/components/schemas/AgentSpecification'
        - type: object
          properties:
            agent_id: { type: string }
            tenant_id: { type: string, format: uuid, nullable: true }
            lifecycle_state:
              type: string
              enum: [DRAFT, SANDBOX, TESTED, EVALUATED, APPROVED, ACTIVE, UPDATED, DEPRECATED, RETIRED]
            evaluation_score: { type: number }
            success_rate: { type: number }
            average_latency_ms: { type: integer }
            average_cost: { type: number }
            version: { type: string }
            parent_agent_id: { type: string, nullable: true }
            active_agent_version_id: { type: string, format: uuid, nullable: true, description: "v1.4 — points at the currently ACTIVE agent_versions row (clause 60)." }
            preferred_provider: { type: string }
            preferred_model: { type: string }
            fallback_models: { type: array, items: { type: string } }
            status: { type: string }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
      description: "v1.4: field set now matches agent_registry columns 1:1 — closes the v1.3 schema/SQL drift finding (clause 69)."
...
    ApprovalRequest:
      type: object
      required: [action, risk_level]
      properties:
        task_id: { type: string }
        agent_id: { type: string }
        action:
          type: string
          description: "Includes AGENT_ACTIVATE for the Factory→Production transition (blueprint clause 45)."
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        reason: { type: string }
    ApprovalRecord:
      allOf:
        - $ref: '#/components/schemas/ApprovalRequest'
        - type: object
          properties:
            request_id: { type: string }
            tenant_id: { type: string, format: uuid }
            requester: { type: string }
            approver: { type: string }
            decision: { type: string, enum: [APPROVED, DENIED, EXPIRED] }
            expires_at: { type: string, format: date-time }
            execution_result: { type: object }
            created_at: { type: string, format: date-time }
            decided_at: { type: string, format: date-time }
...
    AgentVersion:
      type: object
      properties:
        agent_version_id: { type: string, format: uuid }
        agent_id: { type: string }
        version: { type: string }
        specification_hash: { type: string }
        prompt_version: { type: string }
        model_policy: { type: object }
        permissions_snapshot: { type: object }
        evaluation_score: { type: number }
        lifecycle_state: { type: string }
...
  responses:
...
    PolicyDenied:
      description: Central Policy Engine returned DENY
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: Resource not found (or not visible to this tenant)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
paths:
...
  /agents:
    post:
      summary: Submit an agent specification to Agent Factory
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentSpecification' }
      responses:
        '202':
          description: Agent creation workflow started (lands in DRAFT, not ACTIVE)
          content: { application/json: { schema: { $ref: '#/components/schemas/AgentRecord' } } }
        '403': { $ref: '#/components/responses/PolicyDenied' }
    get:
      summary: List agents visible to the caller's tenant
      parameters:
        - { name: lifecycle_state, in: query, schema: { type: string } }
        - { name: department, in: query, schema: { type: string } }
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AgentRecord' } }
  /agents/{agent_id}:
    get:
      summary: Get agent registry record
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Agent record
          content: { application/json: { schema: { $ref: '#/components/schemas/AgentRecord' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /agents/{agent_id}/messages:
    # NOTE (v1.4 integrity fix): this path was declared twice in earlier
    # drafts of this document (once for GET, once for POST). Duplicate YAML
    # mapping keys are undefined behavior — most parsers (js-yaml, PyYAML)
    # silently keep only the LAST occurrence, which would have made the GET
    # operation below disappear entirely from any spec-driven tooling
    # without a single visible error. Merged into one path item here; both
    # operations are implemented in services/control-plane-api.
    get:
      summary: List agent-to-agent messages sent to/from this agent
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        '200':
          description: Message list
          content: { application/json: { schema: { type: array, items: { type: object } } } }
    post:
      summary: Send a typed agent-to-agent message (idempotency-keyed — clause 63)
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiver_agent_id, message_type]
              properties:
                receiver_agent_id: { type: string }
                message_type: { type: string }
                input_payload: { type: object }
      responses:
        '202': { description: Message queued for delivery }
...
  /approvals:
    post:
      summary: Create an approval request (includes AGENT_ACTIVATE)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalRequest' }
      responses:
        '201':
          description: Approval request created
          content: { application/json: { schema: { $ref: '#/components/schemas/ApprovalRecord' } } }
    get:
      summary: List pending approvals for the caller's tenant
      parameters:
        - { name: risk_level, in: query, schema: { type: string, enum: [GREEN, YELLOW, RED] } }
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ApprovalRecord' } }
  /approvals/{request_id}/decision:
    post:
      summary: Record an approval decision (management/Owner action)
      parameters:
        - { name: request_id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [APPROVED, DENIED] }
                reason: { type: string }
      responses:
        '200':
          description: Updated approval record
          content: { application/json: { schema: { $ref: '#/components/schemas/ApprovalRecord' } } }
...
  /agents/{agent_id}/versions:
    post:
      summary: Submit a new agent version (does not activate it)
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string } }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersion' } } }
      responses:
        '201': { description: Version created, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersion' } } } }
    get:
      summary: List all versions of an agent
      parameters: [ { name: agent_id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Version list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AgentVersion' } } } } }
  /agents/{agent_id}/versions/{agent_version_id}/activate:
    post:
      summary: Activate a specific agent version into production (requires prior AGENT_ACTIVATE approval — clause 45/60)
      parameters:
        - { name: agent_id, in: path, required: true, schema: { type: string } }
        - { name: agent_version_id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Version activated }
        '403': { $ref: '#/components/responses/PolicyDenied' }
...